Place calc buttons in a regular calculator layout
Buttons are currently roughly in the same layout as a regular
calculator. Still needs some refining and to add a backspace
button.

Currently numbers and operators are grouped into different
sections. I did this for accessibility reasons, but it may not be
needed or be too complicated. This same layout can be made without
this grouping and all the subgrids, but involves mixing the two
sections and using line placement in a way that might not flow
well. A little more research is probably required!