Update equals function
Now calculates as it should if operator is hit before the equals
at the end of a calculation. Would be better if it functioned as
a typical calculator and pressing operator instead of equals at
end continued function instead of hard stopping it.

How complex I make the calculator depends on how much further I
feel I should take it, or if it's time to move on to the next
project.